[logging] Brand new subsystem...
...This subsystem allows AltaCore to communicate compilation messages properly to whoever's using it.
The logger has 4 message levels: verbose, information, warning, and error. Most are communicated through the logger, but messages may also be thrown when an error that cannot be recovered from internally occurs.
In addition, a warning is now logged when a line starts with a parenthesis which informs the user that this could be misinterpreted as a function call.