[det, ast] Add a wrapper scope for ranged for loops
The wrapper scope defines the lifetime for the counter and the start and (if applicable) end tests. It is different from the loop scope, because the wrapper scope is a one-and-done scope for the entire lifetime of the loop, while the regular loop scope is destroyed and recreated on every iteration of the loop.