[modules] Significantly improve the search algorithm
We now check for `main.alta` if a directory is the target of the import and (this is important!): we don't search parent directories looking for `package.alta.yaml`! I don't know why we used to do that, but that's so wrong that I'm amazed it hasn't bitten me in the ass while developing!