Various fixes and optimizations for type cast paths
Basically, we want to reduce the number of References and Dereferences
we add to the path.

However, in one case, we do want to prefer Dereferencing over From/To
casts: if Dereferencing gives us the exact type we're looking for, prefer
that over doing a From/To cast. This is most obvious with Boxes in the
standard library.